+++ L3 Ateliers POO/Atelier2/util/Secretaire.java
@@ -16,7 +16,7 @@ public class Secretaire extends Employe{
      * @param dateEmbauche la date d'embauche de l'employé
      */
     protected Secretaire(String leNom, String lePrenom, GregorianCalendar laDate,
-    					 Adresse lAdresse, float salaire, GregorianCalendar dateEmbauche) {
+    					 Adresse lAdresse, double salaire, GregorianCalendar dateEmbauche) {
         super(leNom, lePrenom, laDate, lAdresse, salaire, dateEmbauche);
     }
     
@@ -37,7 +37,7 @@ public class Secretaire extends Employe{
      */
     protected Secretaire(String leNom, String lePrenom, int j, int m, int a,
                       int numero, String rue, String code_postal, String ville,
-                      float salaire, int jEmbauche, int mEmbauche, int aEmbauche) {
+                      double salaire, int jEmbauche, int mEmbauche, int aEmbauche) {
         this(leNom, lePrenom, new GregorianCalendar(a,m,j), new Adresse(numero,rue,code_postal,ville),
              salaire, new GregorianCalendar(aEmbauche,mEmbauche,jEmbauche));
     }
@@ -61,7 +61,7 @@ public class Secretaire extends Employe{
     public static Secretaire createSecretaire(String leNom, String lePrenom,
     										  int jNaissance, int mNaissance, int aNaissance,
     										  int numero, String rue, String code_postal, String ville,
-    										  float salaire, int jEmbauche, int mEmbauche, int aEmbauche) {
+    										  double salaire, int jEmbauche, int mEmbauche, int aEmbauche) {
         Secretaire se = null;
         if(salaire > 0) {
             se = new Secretaire(leNom, lePrenom, jNaissance, mNaissance, aNaissance,
@@ -90,7 +90,7 @@ public class Secretaire extends Employe{
 	 * avec un bonus pour chaque manager géré
 	 * @param pourcentage Pourcentage d'augementation du salaire (sans bonus)
 	 */
-	public void augmenterLeSalaire(float pourcentage) {
+	public void augmenterLeSalaire(double pourcentage) {
 		if (pourcentage > 0) {
 			pourcentage += this.getNbManagers() * BONUSSALAIRE;
 			double salaire = getSalaire();
